+++ src/components/core/InputBox/InputBox.module.scss
@@ -1,4 +1,4 @@
-@import '../../variables';
+@import '../../../variables';
 
 .container {
   position: relative;
